Fast restart on busy parents (#19964)
Prepopulate UUIDs to MRG to improve startup times.

- [x] Pre-populate UUIDs to MRG on start (from SQLite)
- [x] Cleanup obsolete UUIDs from MRG once finished

This improves startup times by 3x to 5x.

---------

(cherry picked from commit 3d0c2b9f410564d43b7db3580c5d49417ac0e46b)
