+++ src/libnetdata/libjudy/judyl-typed.h
@@ -13,54 +13,63 @@
         Pvoid_t judyl;                                                           \
     } NAME##_JudyLSet;                                                           \
                                                                                  \
-    static inline void NAME##_INIT(NAME##_JudyLSet *set) {                       \
+    ALWAYS_INLINE                                                                \
+    static void NAME##_INIT(NAME##_JudyLSet *set) {                              \
         set->judyl = NULL;                                                       \
     }                                                                            \
                                                                                  \
-    static inline bool NAME##_SET(NAME##_JudyLSet *set, Word_t index, TYPE value) { \
+    ALWAYS_INLINE                                                                \
+    static bool NAME##_SET(NAME##_JudyLSet *set, Word_t index, TYPE value) { \
         Pvoid_t *pValue = JudyLIns(&set->judyl, index, PJE0);                    \
         if (pValue == PJERR) return false;                                       \
         *pValue = (void *)PACK_MACRO(value);                                     \
         return true;                                                             \
     }                                                                            \
                                                                                  \
-    static inline TYPE NAME##_GET(NAME##_JudyLSet *set, Word_t index) {          \
+    ALWAYS_INLINE                                                                \
+    static TYPE NAME##_GET(NAME##_JudyLSet *set, Word_t index) {                 \
         Pvoid_t *pValue = JudyLGet(set->judyl, index, PJE0);                     \
         return (pValue != NULL) ? (TYPE)UNPACK_MACRO(*pValue) : (TYPE){0};       \
     }                                                                            \
                                                                                  \
-    static inline bool NAME##_DEL(NAME##_JudyLSet *set, Word_t index) {          \
+    ALWAYS_INLINE                                                                \
+    static bool NAME##_DEL(NAME##_JudyLSet *set, Word_t index) {                 \
         return JudyLDel(&set->judyl, index, PJE0) == 1;                          \
     }                                                                            \
                                                                                  \
-    static inline TYPE NAME##_FIRST(NAME##_JudyLSet *set, Word_t *index) {       \
+    ALWAYS_INLINE                                                                \
+    static TYPE NAME##_FIRST(NAME##_JudyLSet *set, Word_t *index) {              \
         Pvoid_t *pValue = JudyLFirst(set->judyl, index, PJE0);                   \
         return (pValue != NULL) ? (TYPE)UNPACK_MACRO(*pValue) : (TYPE){0};       \
     }                                                                            \
                                                                                  \
-    static inline TYPE NAME##_NEXT(NAME##_JudyLSet *set, Word_t *index) {        \
+    ALWAYS_INLINE                                                                \
+    static TYPE NAME##_NEXT(NAME##_JudyLSet *set, Word_t *index) {               \
         Pvoid_t *pValue = JudyLNext(set->judyl, index, PJE0);                    \
         return (pValue != NULL) ? (TYPE)UNPACK_MACRO(*pValue) : (TYPE){0};       \
     }                                                                            \
                                                                                  \
-    static inline TYPE NAME##_LAST(NAME##_JudyLSet *set, Word_t *index) {        \
+    ALWAYS_INLINE                                                                \
+    static TYPE NAME##_LAST(NAME##_JudyLSet *set, Word_t *index) {               \
         Pvoid_t *pValue = JudyLLast(set->judyl, index, PJE0);                    \
         return (pValue != NULL) ? (TYPE)UNPACK_MACRO(*pValue) : (TYPE){0};       \
     }                                                                            \
                                                                                  \
-    static inline TYPE NAME##_PREV(NAME##_JudyLSet *set, Word_t *index) {        \
+    ALWAYS_INLINE                                                                \
+    static TYPE NAME##_PREV(NAME##_JudyLSet *set, Word_t *index) {               \
         Pvoid_t *pValue = JudyLPrev(set->judyl, index, PJE0);                    \
         return (pValue != NULL) ? (TYPE)UNPACK_MACRO(*pValue) : (TYPE){0};       \
     }                                                                            \
                                                                                  \
-    static inline void NAME##_FREE(NAME##_JudyLSet *set, void (*callback)(Word_t, TYPE)) { \
+    ALWAYS_INLINE                                                                \
+    static void NAME##_FREE(NAME##_JudyLSet *set, void (*callback)(Word_t, TYPE, void *), void *data) { \
         Word_t index = 0;                                                        \
         Pvoid_t *pValue;                                                         \
         if (callback) {                                                          \
             for (pValue = JudyLFirst(set->judyl, &index, PJE0);                  \
                  pValue != NULL;                                                 \
                  pValue = JudyLNext(set->judyl, &index, PJE0)) {                 \
-                callback(index, (TYPE)UNPACK_MACRO(*pValue));                    \
+                callback(index, (TYPE)UNPACK_MACRO(*pValue), data);              \
             }                                                                    \
         }                                                                        \
         JudyLFreeArray(&set->judyl, PJE0);                                       \
